Split out logging code from util.c
The functions that log data to a file make use of parts of main.h
and (annoyingly) maps.h.

In my effort to reduce the interdependence between util.c and other
parts of the code (so that functions here could potentially be used in
unit tests), this is an issue.

With this commit, I can remove the include of main.h and maps.h from
util.c.

As such, I'm calling it a day on this effort, and will submit my work
so far as a pull request.  There will be more pull requests related to
