+++ src/messages_gui.c
@@ -39,6 +39,7 @@
 #include "lang.h"
 #include "mutex_utils.h"
 #include "xa_config.h"
+#include "log_utils.h"
 
 // Must be last include file
 #include "leak_detection.h"
